+++ client/src/components/UI/organisms/index.ts
@@ -1,2 +1,3 @@
 export { ModalPopup } from './ModalPopup/ModalPopup';
 export { SearchResults } from './SearchResults/SearchResults';
+export { Header } from './Header/Header';
